fix types for id

+++ src/main/kotlin/data/remote/models/ReversedCountry.kt
@@ -11,9 +11,9 @@ data class ReversedCountry(
     val licence: String,
     val lon: String,
     val name: String,
-    val osm_id: Int,
+    val osm_id: Long,
     val osm_type: String,
-    val place_id: Int,
+    val place_id: Long,
     val place_rank: Int,
     val type: String
 )
